Moved contactListItem to sperate component as part of seperation of concern
