+++ src/Components/ContactBook.js
@@ -4,40 +4,13 @@ import {
 	Paper,
 	Grid,
 	Typography,
-	Avatar,
 	List,
-	ListItem,
-	ListItemAvatar,
-	ListItemText,
 } from '@material-ui/core'
-import ContactPhoneIcon from '@material-ui/icons/ContactPhone'
+import ContactListItem from './ContactListItem'
 
 const useStyles = makeStyles((theme) => ({
-	contactList: {
-		width: '350px',
-		maxHeight: '500px',
-		scrollable: true,
-	},
-	avatar: {
-		backgroundColor: theme.palette.secondary.main,
-	},
+
 }))
-
-const ContactListItem = (props) => {
-    const classes = useStyles()
-    const contact = props.contact
-
-	return (
-		<ListItem>
-			<ListItemAvatar>
-				<Avatar className={classes.avatar}>
-					<ContactPhoneIcon />
-				</Avatar>
-			</ListItemAvatar>
-			<ListItemText primary={contact.name} secondary={contact.job_title} />
-		</ListItem>
-	)
-}
 
 export default function ContactBook(props) {
 	const classes = useStyles()
